+++ CHANGELOG.md
@@ -3,6 +3,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
+### 🔄 Changed
+
+# [5.1.0](https://github.com/GetStream/stream-chat-swift/releases/tag/5.1.0)
+_April 23, 2026_
+
 ## StreamChatUI
 ### 🔄 Changed
 - `CDNRequester` is now passed in the constructor of `StreamMediaLoader` instead of `Components` [#4070](https://github.com/GetStream/stream-chat-swift/pull/4070)